LPS-199325 Move duplicated logic in ActionEditableElementMapper and DisplayPageInfoItemFieldSetProviderImpl to DisplayPageURLProviderImpl
