+++ modules/apps/layout/layout-page-template-service/src/main/java/com/liferay/layout/page/template/internal/info/item/provider/DisplayPageInfoItemFieldSetProviderImpl.java
@@ -5,10 +5,6 @@
 
 package com.liferay.layout.page.template.internal.info.item.provider;
 
-import com.liferay.asset.display.page.portlet.AssetDisplayPageFriendlyURLProvider;
-import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
-import com.liferay.asset.kernel.model.AssetRenderer;
-import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.info.field.InfoField;
 import com.liferay.info.field.InfoFieldSet;
 import com.liferay.info.field.InfoFieldSetEntry;
@@ -16,36 +12,28 @@ import com.liferay.info.field.InfoFieldValue;
 import com.liferay.info.field.type.DisplayPageInfoFieldType;
 import com.liferay.info.field.type.InfoFieldType;
 import com.liferay.info.field.type.URLInfoFieldType;
-import com.liferay.info.item.ClassPKInfoItemIdentifier;
-import com.liferay.info.item.ERCInfoItemIdentifier;
-import com.liferay.info.item.InfoItemIdentifier;
 import com.liferay.info.item.InfoItemReference;
 import com.liferay.info.localized.InfoLocalizedValue;
 import com.liferay.info.localized.bundle.FunctionInfoLocalizedValue;
 import com.liferay.info.type.WebURL;
 import com.liferay.layout.page.template.constants.LayoutPageTemplateEntryTypeConstants;
 import com.liferay.layout.page.template.info.item.provider.DisplayPageInfoItemFieldSetProvider;
+import com.liferay.layout.page.template.info.item.provider.DisplayPageURLProvider;
 import com.liferay.layout.page.template.model.LayoutPageTemplateEntry;
 import com.liferay.layout.page.template.service.LayoutPageTemplateEntryService;
-import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.Layout;
-import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Portal;
-import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Locale;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -100,7 +88,8 @@ public class DisplayPageInfoItemFieldSetProviderImpl
 				).labelInfoLocalizedValue(
 					InfoLocalizedValue.localize(getClass(), "default")
 				).build(),
-				_getDefaultDisplayPageURL(infoItemReference, themeDisplay)));
+				_displayPageURLProvider.getDefaultURL(
+					infoItemReference, themeDisplay)));
 
 		List<LayoutPageTemplateEntry> layoutPageTemplateEntries =
 			_layoutPageTemplateEntryService.getLayoutPageTemplateEntries(
@@ -135,7 +124,7 @@ public class DisplayPageInfoItemFieldSetProviderImpl
 
 							try {
 								webURL = new WebURL(
-									_getDisplayPageTemplateURL(
+									_displayPageURLProvider.getURL(
 										infoItemReference,
 										layoutPageTemplateEntry, locale,
 										themeDisplay));
@@ -157,55 +146,6 @@ public class DisplayPageInfoItemFieldSetProviderImpl
 		return infoFieldValues;
 	}
 
-	private String _getDefaultDisplayPageURL(
-			InfoItemReference infoItemReference, ThemeDisplay themeDisplay)
-		throws Exception {
-
-		AssetRendererFactory<?> assetRendererFactory =
-			AssetRendererFactoryRegistryUtil.getAssetRendererFactoryByClassName(
-				infoItemReference.getClassName());
-
-		if (assetRendererFactory == null) {
-			return _assetDisplayPageFriendlyURLProvider.getFriendlyURL(
-				infoItemReference, themeDisplay);
-		}
-
-		try {
-			AssetRenderer<?> assetRenderer = null;
-
-			if (infoItemReference.getInfoItemIdentifier() instanceof
-					ClassPKInfoItemIdentifier) {
-
-				ClassPKInfoItemIdentifier classPKInfoItemIdentifier =
-					(ClassPKInfoItemIdentifier)
-						infoItemReference.getInfoItemIdentifier();
-
-				assetRenderer = assetRendererFactory.getAssetRenderer(
-					classPKInfoItemIdentifier.getClassPK());
-			}
-
-			if (assetRenderer == null) {
-				return _assetDisplayPageFriendlyURLProvider.getFriendlyURL(
-					infoItemReference, themeDisplay);
-			}
-
-			String viewInContextURL = assetRenderer.getURLViewInContext(
-				themeDisplay, StringPool.BLANK);
-
-			if (Validator.isNotNull(viewInContextURL)) {
-				return viewInContextURL;
-			}
-		}
-		catch (Exception exception) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(exception);
-			}
-		}
-
-		return _assetDisplayPageFriendlyURLProvider.getFriendlyURL(
-			infoItemReference, themeDisplay);
-	}
-
 	private InfoField<InfoFieldType> _getDefaultDisplayPageURLInfoField(
 		String namespace) {
 
@@ -222,25 +162,6 @@ public class DisplayPageInfoItemFieldSetProviderImpl
 
 	private InfoFieldType _getDisplayPageInfoFieldType() {
 		return DisplayPageInfoFieldType.INSTANCE;
-	}
-
-	private String _getDisplayPageTemplateURL(
-			InfoItemReference infoItemReference,
-			LayoutPageTemplateEntry layoutPageTemplateEntry, Locale locale,
-			ThemeDisplay themeDisplay)
-		throws PortalException {
-
-		Layout layout = _layoutLocalService.fetchLayout(
-			layoutPageTemplateEntry.getPlid());
-
-		Group group = themeDisplay.getScopeGroup();
-
-		return StringBundler.concat(
-			_portal.getGroupFriendlyURL(
-				group.getPublicLayoutSet(), themeDisplay, false, false),
-			"/e", layout.getFriendlyURL(locale), StringPool.SLASH,
-			_portal.getClassNameId(infoItemReference.getClassName()),
-			StringPool.SLASH, _getInfoItemIdentifier(infoItemReference));
 	}
 
 	private List<InfoFieldSetEntry> _getInfoFieldSetEntries(
@@ -279,27 +200,6 @@ public class DisplayPageInfoItemFieldSetProviderImpl
 		return infoFieldSetEntries;
 	}
 
-	private String _getInfoItemIdentifier(InfoItemReference infoItemReference) {
-		InfoItemIdentifier infoItemIdentifier =
-			infoItemReference.getInfoItemIdentifier();
-
-		if (infoItemIdentifier instanceof ClassPKInfoItemIdentifier) {
-			ClassPKInfoItemIdentifier classPKInfoItemIdentifier =
-				(ClassPKInfoItemIdentifier)infoItemIdentifier;
-
-			return String.valueOf(classPKInfoItemIdentifier.getClassPK());
-		}
-
-		if (infoItemIdentifier instanceof ERCInfoItemIdentifier) {
-			ERCInfoItemIdentifier ercInfoItemIdentifier =
-				(ERCInfoItemIdentifier)infoItemIdentifier;
-
-			return ercInfoItemIdentifier.getExternalReferenceCode();
-		}
-
-		return StringPool.BLANK;
-	}
-
 	private String _getUniqueId(String layoutPageTemplateEntryKey) {
 		return LayoutPageTemplateEntry.class.getSimpleName() +
 			StringPool.UNDERLINE + layoutPageTemplateEntryKey;
@@ -309,11 +209,7 @@ public class DisplayPageInfoItemFieldSetProviderImpl
 		DisplayPageInfoItemFieldSetProviderImpl.class);
 
 	@Reference
-	private AssetDisplayPageFriendlyURLProvider
-		_assetDisplayPageFriendlyURLProvider;
-
-	@Reference
-	private LayoutLocalService _layoutLocalService;
+	private DisplayPageURLProvider _displayPageURLProvider;
 
 	@Reference
 	private LayoutPageTemplateEntryService _layoutPageTemplateEntryService;
